src/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.4.0
autopep8==1.6.0
azure-core==1.26.4
azure-identity==1.13.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.1
chainlit==0.6.1
charset-normalizer==2.0.7
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.1.6
clickhouse-connect==0.5.24
cmake==3.26.3
coloredlogs==15.0.1
cryptography==40.0.2
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dotenv-python==0.0.1
duckdb==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.97.0
fastapi-socketio==0.0.10
filelock==3.12.0
filetype==1.2.0
Flask==2.0.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.3
importlib-metadata==6.0.1
importlib-resources==6.0.0
itsdangerous==2.0.1
jc==1.23.1
Jinja2==3.0.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.244
langsmith==0.0.14
lit==16.0.3
llama-index==0.7.13
lxml==4.9.3
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.0.9
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pandas-stubs==1.2.0.35
Pillow==9.5.0
pipdeptree==2.7.1
portalocker==2.7.0
posthog==3.0.1
prisma==0.9.1
protobuf==3.20.3
pulsar-client==3.2.0
pyarrow==12.0.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pypdf==3.12.2
PyPDF2==3.0.1
PyPika==0.48.9
pyrsistent==0.19.3
pysqlite3-binary==0.5.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-socketio==5.8.0
pytz==2021.3
PyYAML==6.0
regex==2023.5.5
requests==2.29.0
rich==13.3.5
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
starlette==0.27.0
streamlit==1.22.0
streamlit-chat==0.0.2.2
sympy==1.12
syncer==2.0.3
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.29.1
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uptrace==1.18.0
urllib3==1.26.7
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.0.2
wrapt==1.15.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
